aarch64: Move AMode computation into ISLE  (#6805)
* Use `Offset32` as `i32` in ISLE

This commit updates the x64 and aarch64 backends to use the `i32`
primitive type in ISLE when working with an `Offset32` instead of a
`u32`. This matches the intended representation of `Offset32` as a type
which is signed internally and represents how offsets on instructions
are often negative too.

This does not actually change any end results of compilation and instead
is intended to be "just" an internal refactoring with fewer casts and
more consistent handling of offsets.

* aarch64: Define the `PairAMode` type in ISLE

This commit moves the definition of the `PairAMode` enum into ISLE
instead of its current Rust-defined location. This is in preparation for
the next commit where all AMode calculations will be moved into ISLE.

* aarch64: Fix a copy/paste typo loading vectors

This commit fixes an assertion that can be tripped in the aarch64
backend where a 64-bit load was accidentally flagged as a 128-bit load.
This was found in future work which ended up tripping the assertion a
bit earlier.

* aarch64: Move AMode computation into ISLE

This commit moves the computation of the `AMode` enum for addressing
from Rust into ISLE. This enables deleting a good deal of Rust code in
favor of (hopefully) more readable ISLE code.

This does not mirror the preexisting logic exactly but instead takes a
different approach for generating the `AMode`. Previously the entire
chain of `iadd`s input into an address were unfolded into 32-bit and
64-bit operations and then those were re-combined as possible into an
`AMode` (possibly emitting `add` instructions. Instead now pattern
matching is used to represent this. The net result is that amodes are
emitted slightly differently here and there in a number of updated test
cases.

I've tried to verify in all test cases that the number of instructions
has not increased and the same logical operation is happening. The exact
`AMode` may differ but at least instruction-wise this shouldn't be a
regression. My hope is that if anything needs changing that can be
represented with updates to the rule precedence in ISLE or various other
special cases.

One part I found a little surprising was that the immediate with a
load/store instruction is not actually used much of the time. I naively
thought that the mid-end optimizations would move iadd immediates into
the load/store immediate but that is not the case. This necessitated two
extra ISLE rules to manually peel off immediates and fold them into the
load/store immediate.

* aarch64: Remove `NarrowValueMode`

This is no longer needed after the prior commit

+++ winch/codegen/src/isa/x64/asm.rs
@@ -225,7 +225,7 @@ impl Assembler {
 
     /// Register-to-memory move.
     pub fn mov_rm(&mut self, src: Reg, base: Reg, disp: u32, size: OperandSize) {
-        let dst = Amode::imm_reg(disp, base.into());
+        let dst = Amode::imm_reg(disp as i32, base.into());
 
         self.emit(Inst::MovRM {
             size: size.into(),
@@ -236,7 +236,7 @@ impl Assembler {
 
     /// Immediate-to-memory move.
     pub fn mov_im(&mut self, src: u64, base: Reg, disp: u32, size: OperandSize) {
-        let dst = Amode::imm_reg(disp, base.into());
+        let dst = Amode::imm_reg(disp as i32, base.into());
         self.emit(Inst::MovImmM {
             size: size.into(),
             simm64: src,
@@ -260,7 +260,7 @@ impl Assembler {
     pub fn mov_mr(&mut self, base: Reg, disp: u32, dst: Reg, size: OperandSize) {
         use OperandSize::S64;
 
-        let amode = Amode::imm_reg(disp, base.into());
+        let amode = Amode::imm_reg(disp as i32, base.into());
         let src = SyntheticAmode::real(amode);
 
         if size == S64 {
